Move all node APIs requiring components of detail out of public header to recompilation of AST-consuming code when AST internals change.
